% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{R87}
\alias{R87}
\title{Table R87}
\format{
A data frame with 2 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{MahMorajeh}:
\itemize{
\item Farsi name: \emph{ماه مراجعه به خانوار}
\item Description: The month of visiting the household
}
}
}
\usage{
R87
}
\description{
{R for Rural} \cr
{87 for 1387}
}
\keyword{datasets}
