% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{R87Data}
\alias{R87Data}
\title{Table R87Data \cr
Including some information on the front and back of the questionnaire in 1387 \cr
شامل برخی اطلاعات رو و پشت جلد پرسشنامه در سال 1387}
\format{
A data frame with 19 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{MahMorajeh}:
\itemize{
\item Farsi name: \emph{ماه مراجعه به خانوار}
\item Description: The month of visiting the household
}
\item \code{NeshaniSarparast}:
\itemize{
\item Farsi name: \emph{نشانی سرپرست خانوار}
\item Description: The address of the head of the household
}
\item \code{NameSarparast}:
\itemize{
\item Farsi name: \emph{نام سرپرست خانوار}
\item Description: The name of the head of the household
}
\item \code{ShoghlSarparast}:
\itemize{
\item Farsi name: \emph{شغل سرپرست خانوار}
\item Description: The job of the head of the household
}
\item \code{NoeKhn}:
\itemize{
\item Farsi name: \emph{نوع خانوار (معمولي ساكن كد 1 گروهي كد 2)}
\item Description: Household type (ordinary resident code 1, group code 2)
}
\item \code{Takmil}:
\itemize{
\item Farsi name: \emph{بخش پ روي جلد پرسشنامه، سوال 19 (بله كد 1 نه كد 2)}
\item Description: Questionnaire for substitute household (Yes code 1, No code 2)
}
\item \code{TakmilDescC}:
\itemize{
\item Farsi name: \emph{كد 7 جدول -3 ساير (با ذكر علت)}
\item Description: (Code 7, table-3, other (with the reason for replacing the household)
}
\item \code{Jaygozin}:
\itemize{
\item Farsi name: \emph{اطلاع پرسش 20 (بله كد ،1 نه كد 2)}
\item Description:
}
\item \code{JaygozinDesc}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{BlkAbdJaygozin}:
\itemize{
\item Farsi name: \emph{شماره بلوك يا آبادي خانوار جايگزين (كه بايد با خانوار نمونه ی اصلي يكسان باشد)}
\item Description: Block or settlement number of the alternative household (which must be the same as the original sample household)
}
\item \code{RadifJaygozin}:
\itemize{
\item Farsi name: \emph{شماره رديف خانوار جايگزين در واحد نمونه گيري اوليه}
\item Description:  Substitute household row number in the primary sampling unit
}
\item \code{fStat}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{fEnt}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{fvfy}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{user1}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{user2}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{user3}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
\item \code{user4}:
\itemize{
\item Farsi name: \emph{اطلاعاتی ثبت نشده است}
\item Description: No description
}
}
}
\usage{
R87Data
}
\description{
{R for Rural} \cr
{87 for 1387}
}
\keyword{datasets}
