% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data_documentation.R
\docType{data}
\name{R87p3S01}
\alias{R87p3S01}
\title{Table R87p3S01 \cr
Household food and tobacco expenditure of rural household in 1387 \cr
هزينه هاي خوراكي و دخانی خانوار روستایی سال 1387}
\format{
A data frame with 7 columns:
\itemize{
\item \code{Address}:
\itemize{
\item Farsi name: \emph{آدرس}
\item Description: Household address
}
\item \code{DYCOL01}:
\itemize{
\item Farsi name: \emph{کد کالا}
\item Description: Bar code
}
\item \code{DYCOL02}:
\itemize{
\item Farsi name: \emph{طریق تهیه}
\item Description: Preparation method
}
\item \code{DYCOL03}:
\itemize{
\item Farsi name: \emph{مقدار (گرم)}
\item Description: Amount(gram)
}
\item \code{DYCOL04}:
\itemize{
\item Farsi name: \emph{مقدار (کیلوگرم)}
\item Description: Amount(kilogram)
}
\item \code{DYCOL05}:
\itemize{
\item Farsi name: \emph{قیمت واحد (ریال)}
\item Description: Unit price (Rials)
}
\item \code{DYCOL06}:
\itemize{
\item Farsi name: \emph{ارزش (ریال)}
\item Description: Value (Rials)
}
}
}
\usage{
R87p3S01
}
\description{
{R for Rural} \cr
{87 for 1387} \cr
{P for Part} \cr
{S for Section}
}
\keyword{datasets}
